HW03 fixed as per Review 1

+++ src/main/java/ru/javawebinar/topjava/service/MealService.java
@@ -1,9 +1,11 @@
 package ru.javawebinar.topjava.service;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.lang.Nullable;
 import org.springframework.stereotype.Service;
 import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
+import ru.javawebinar.topjava.util.exception.NotFoundException;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -14,6 +16,8 @@ import static ru.javawebinar.topjava.util.ValidationUtil.checkNotFoundWithId;
 
 @Service
 public class MealService {
+    @Autowired
+    private UserService userService;
 
     private final MealRepository repository;
 
@@ -26,6 +30,9 @@ public class MealService {
     }
 
     public void delete(int id, int userId) {
+        if (get(id, userId) == null) {
+            throw (new NotFoundException("Deleting non-existing Meal"));
+        }
         checkNotFoundWithId(repository.delete(id, userId), id);
     }
 
@@ -38,10 +45,16 @@ public class MealService {
     }
 
     public void update(Meal meal, int userId) {
+        if (get(meal.getId(), userId) == null) {
+            throw (new NotFoundException("Meal updating data is inconsistent"));
+        }
         checkNotFoundWithId(repository.save(meal, userId), meal.getId());
     }
 
     public Meal create(Meal meal, int userId) {
+        if (userService.get(userId) == null) {
+            throw new NotFoundException("Attempt to create meal for non-existing user");
+        }
         return repository.save(meal, userId);
     }
 }
